move ByKeySpace into separate file
